// apps/desktop/src/generated/ipc/PublicError.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ErrorSubsystem } from "./ErrorSubsystem";
import type { UserAction } from "./UserAction";

export type PublicError = { code: string, subsystem: ErrorSubsystem, operation_id: string | null, retryable: boolean, user_action: UserAction, safe_message_key: string, safe_parameters: { [key in string]?: string }, diagnostic_id: string, };